Clean up duplicate test classes (#427)
* Clean up duplicate test classes

* Minor update

* Fix test failure

* Fix test failure

* minor updates

* Fix build failure
