+++ src/Microsoft.Azure.SignalR.AspNet/ClientConnections/ClientConnectionManager.cs
@@ -22,11 +22,10 @@ namespace Microsoft.Azure.SignalR.AspNet
 
         private readonly ConcurrentDictionary<string, IServiceConnection> _clientConnections = new ConcurrentDictionary<string, IServiceConnection>();
 
-        public ClientConnectionManager(HubConfiguration configuration)
+        public ClientConnectionManager(HubConfiguration configuration, ILoggerFactory loggerFactory)
         {
             _configuration = configuration;
-            var loggerFactory = configuration.Resolver.Resolve<ILoggerFactory>() ?? NullLoggerFactory.Instance;
-            _logger = loggerFactory.CreateLogger<ClientConnectionManager>();
+            _logger = loggerFactory?.CreateLogger<ClientConnectionManager>() ?? NullLogger<ClientConnectionManager>.Instance;
         }
 
         public IServiceTransport CreateConnection(OpenConnectionMessage message, IServiceConnection serviceConnection)
