+++ test/Microsoft.Azure.SignalR.AspNet.Tests/ClientConnectionManagerTests.cs
@@ -20,11 +20,14 @@ namespace Microsoft.Azure.SignalR.AspNet.Tests
 
         public ClientConnectionManagerTests()
         {
-            var hubConfig = new HubConfiguration();
+            var hubConfig = new HubConfiguration
+            {
+                Resolver = new DefaultDependencyResolver()
+            };
             var transport = new AzureTransportManager(hubConfig.Resolver);
             hubConfig.Resolver.Register(typeof(ITransportManager), () => transport);
 
-            _clientConnectionManager = new ClientConnectionManager(hubConfig);
+            _clientConnectionManager = new ClientConnectionManager(hubConfig, null);
         }
 
         [Theory]
