+++ test/Microsoft.Azure.SignalR.AspNet.Tests/ServiceConnectionTests.cs
@@ -2,16 +2,15 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System;
-using System.Collections.Generic;
 using System.Security.Claims;
 using System.Text;
 using System.Threading.Tasks;
 using Microsoft.AspNet.SignalR;
 using Microsoft.AspNet.SignalR.Transports;
+using Microsoft.Azure.SignalR.AspNet.Tests.TestHubs;
 using Microsoft.Azure.SignalR.Protocol;
 using Microsoft.Azure.SignalR.TestsCommon;
 using Microsoft.Extensions.Logging;
-using Owin;
 using Xunit;
 using Xunit.Abstractions;
 
@@ -20,17 +19,9 @@ namespace Microsoft.Azure.SignalR.AspNet.Tests
     public partial class ServiceConnectionTests : VerifiableLoggedTest
     {
         private const string ConnectionString = "Endpoint=http://localhost;AccessKey=ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789;";
-        private static readonly ServiceProtocol Protocol = new ServiceProtocol();
-
-        private readonly TestConnectionManager _clientConnectionManager;
 
         public ServiceConnectionTests(ITestOutputHelper output) : base(output)
         {
-            var hubConfig = new HubConfiguration();
-            var transport = new AzureTransportManager(hubConfig.Resolver);
-            hubConfig.Resolver.Register(typeof(ITransportManager), () => transport);
-
-            _clientConnectionManager = new TestConnectionManager();
         }
 
         [Fact]
@@ -39,7 +30,12 @@ namespace Microsoft.Azure.SignalR.AspNet.Tests
             int count = 0;
             using (StartVerifiableLog(out var loggerFactory, LogLevel.Debug))
             {
-                using (var proxy = new ServiceConnectionProxy(_clientConnectionManager, loggerFactory: loggerFactory))
+                var hubConfig = Utility.GetTestHubConfig(loggerFactory);
+                var atm = new AzureTransportManager(hubConfig.Resolver);
+                hubConfig.Resolver.Register(typeof(ITransportManager), () => atm);
+
+                var clientConnectionManager = new TestClientConnectionManager();
+                using (var proxy = new TestServiceConnectionProxy(clientConnectionManager, loggerFactory: loggerFactory))
                 {
                     // start the server connection
                     await proxy.StartServiceAsync().OrTimeout();
@@ -55,7 +51,7 @@ namespace Microsoft.Azure.SignalR.AspNet.Tests
                     while (count < 1000)
                     {
                         task = proxy.WaitForApplicationMessageAsync(clientConnection).OrTimeout();
-                        await proxy.WriteMessageAsync(new ConnectionDataMessage(clientConnection, GetPayload("Hello World")));
+                        await proxy.WriteMessageAsync(new ConnectionDataMessage(clientConnection, "Hello World".GenerateSingleFrameBuffer()));
                         await task;
                         count++;
                     }
@@ -65,7 +61,7 @@ namespace Microsoft.Azure.SignalR.AspNet.Tests
                     await task;
 
                     // Validate in transport for 1000 data messages.
-                    _clientConnectionManager.CurrentTransports.TryGetValue(clientConnection, out var transport);
+                    clientConnectionManager.CurrentTransports.TryGetValue(clientConnection, out var transport);
                     Assert.NotNull(transport);
                     await transport.WaitOnDisconnected().OrTimeout();
                     Assert.Equal(transport.MessageCount, count);
@@ -78,17 +74,14 @@ namespace Microsoft.Azure.SignalR.AspNet.Tests
         {
             using (StartVerifiableLog(out var loggerFactory, LogLevel.Debug))
             {
-                var hubConfig = new HubConfiguration();
-                hubConfig.Resolver = new DefaultDependencyResolver();
-                var scm = new TestServiceConnectionManager();
+                var hubConfig = Utility.GetActualHubConfig(loggerFactory);
+
+                var scm = new TestServiceConnectionHandler();
                 hubConfig.Resolver.Register(typeof(IServiceConnectionManager), () => scm);
-
-                var ccm = new ClientConnectionManager(hubConfig);
+                var ccm = new ClientConnectionManager(hubConfig, loggerFactory);
                 hubConfig.Resolver.Register(typeof(IClientConnectionManager), () => ccm);
-
                 DispatcherHelper.PrepareAndGetDispatcher(new TestAppBuilder(), hubConfig, new ServiceOptions { ConnectionString = ConnectionString }, "app1", loggerFactory);
-
-                using (var proxy = new ServiceConnectionProxy(ccm, loggerFactory: loggerFactory))
+                using (var proxy = new TestServiceConnectionProxy(ccm, loggerFactory: loggerFactory))
                 {
                     // start the server connection
                     await proxy.StartServiceAsync().OrTimeout();
@@ -103,17 +96,19 @@ namespace Microsoft.Azure.SignalR.AspNet.Tests
                     // group message goes into the manager
                     // make sure the tcs is called before writing message
                     var jgTask = scm.WaitForTransportOutputMessageAsync(typeof(JoinGroupMessage)).OrTimeout();
+
                     var gbTask = scm.WaitForTransportOutputMessageAsync(typeof(GroupBroadcastDataMessage)).OrTimeout();
 
                     await proxy.WriteMessageAsync(new ConnectionDataMessage(clientConnection, Encoding.UTF8.GetBytes("{\"H\":\"chat\",\"M\":\"JoinGroup\",\"A\":[\"user1\",\"message1\"],\"I\":1}")));
                     
                     await jgTask;
                     await gbTask;
-                    
+
                     var lgTask = scm.WaitForTransportOutputMessageAsync(typeof(LeaveGroupMessage)).OrTimeout();
                     gbTask = scm.WaitForTransportOutputMessageAsync(typeof(GroupBroadcastDataMessage)).OrTimeout();
 
                     await proxy.WriteMessageAsync(new ConnectionDataMessage(clientConnection, Encoding.UTF8.GetBytes("{\"H\":\"chat\",\"M\":\"LeaveGroup\",\"A\":[\"user1\",\"message1\"],\"I\":1}")));
+
                     await lgTask;
                     await gbTask;
 
@@ -122,34 +117,6 @@ namespace Microsoft.Azure.SignalR.AspNet.Tests
                     await dTask;
                 }
             }
-        }
-
-        private sealed class TestAppBuilder : IAppBuilder
-        {
-            public IDictionary<string, object> Properties { get; } = new Dictionary<string, object>()
-            {
-                ["builder.AddSignatureConversion"] = new Action<Delegate>(e => { })
-            };
-
-            public object Build(Type returnType)
-            {
-                return null;
-            }
-
-            public IAppBuilder New()
-            {
-                return null;
-            }
-
-            public IAppBuilder Use(object middleware, params object[] args)
-            {
-                return this;
-            }
-        }
-
-        private ReadOnlyMemory<byte> GetPayload(string message)
-        {
-            return Protocol.GetMessageBytes(new ConnectionDataMessage(string.Empty, System.Text.Encoding.UTF8.GetBytes(message)));
         }
     }
 }
