+++ test/Microsoft.Azure.SignalR.AspNet.Tests/TestClasses/TestServiceMessageHandler.cs
@@ -1,0 +1,20 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System.Threading.Tasks;
+using Microsoft.Azure.SignalR.Protocol;
+
+namespace Microsoft.Azure.SignalR.AspNet.Tests
+{
+    internal sealed class TestServiceMessageHandler : IServiceMessageHandler
+    {
+        public TestServiceMessageHandler()
+        {
+        }
+
+        public Task HandlePingAsync(PingMessage pingMessage)
+        {
+            return Task.CompletedTask;
+        }
+    }
+}
