+++ test/Microsoft.Azure.SignalR.AspNet.Tests/TestClasses/TestTransport.cs
